ArC: introduce CurrentManagedContext
- share the logic used in RequestContext and WebSocketSessionContext
