+++ independent-projects/arc/runtime/src/main/java/io/quarkus/arc/impl/CurrentManagedContext.java
@@ -1,0 +1,270 @@
+package io.quarkus.arc.impl;
+
+import java.lang.invoke.MethodHandles;
+import java.lang.invoke.VarHandle;
+import java.util.Map;
+import java.util.Objects;
+import java.util.function.Consumer;
+import java.util.function.Function;
+import java.util.function.Supplier;
+import java.util.stream.Collectors;
+
+import jakarta.enterprise.context.ContextNotActiveException;
+import jakarta.enterprise.context.spi.Contextual;
+import jakarta.enterprise.context.spi.CreationalContext;
+
+import org.jboss.logging.Logger;
+
+import io.quarkus.arc.ContextInstanceHandle;
+import io.quarkus.arc.CurrentContext;
+import io.quarkus.arc.InjectableBean;
+import io.quarkus.arc.ManagedContext;
+
+/**
+ * A managed context backed by the {@link CurrentContext}.
+ */
+public abstract class CurrentManagedContext implements ManagedContext {
+
+    private static final Logger LOG = Logger.getLogger(CurrentManagedContext.class);
+
+    private final CurrentContext<CurrentContextState> currentContext;
+
+    private final Supplier<ContextInstances> contextInstances;
+
+    private final Consumer<Object> initializedNotifier;
+    private final Consumer<Object> beforeDestroyedNotifier;
+    private final Consumer<Object> destroyedNotifier;
+
+    protected CurrentManagedContext(CurrentContext<CurrentContextState> currentContext,
+            Supplier<ContextInstances> contextInstances, Consumer<Object> initializedNotifier,
+            Consumer<Object> beforeDestroyedNotifier, Consumer<Object> destroyedNotifier) {
+        this.currentContext = currentContext;
+        this.contextInstances = contextInstances;
+        this.initializedNotifier = initializedNotifier;
+        this.beforeDestroyedNotifier = beforeDestroyedNotifier;
+        this.destroyedNotifier = destroyedNotifier;
+    }
+
+    @Override
+    public ContextState getState() {
+        CurrentContextState state = currentState();
+        if (state == null) {
+            throw notActive();
+        }
+        return state;
+    }
+
+    @Override
+    public ContextState activate(ContextState initialState) {
+        if (traceLog().isTraceEnabled()) {
+            traceActivate(initialState);
+        }
+        if (initialState == null) {
+            CurrentContextState state = initializeContextState();
+            currentContext.set(state);
+            return state;
+        } else {
+            if (initialState instanceof CurrentContextState) {
+                currentContext.set((CurrentContextState) initialState);
+                return initialState;
+            } else {
+                throw new IllegalArgumentException("Invalid initial state: " + initialState.getClass().getName());
+            }
+        }
+    }
+
+    @Override
+    public void deactivate() {
+        if (traceLog().isTraceEnabled()) {
+            traceDeactivate();
+        }
+        currentContext.remove();
+    }
+
+    @SuppressWarnings("unchecked")
+    @Override
+    public <T> T getIfActive(Contextual<T> contextual, Function<Contextual<T>, CreationalContext<T>> creationalContextFun) {
+        Objects.requireNonNull(contextual, "Contextual must not be null");
+        Objects.requireNonNull(creationalContextFun, "CreationalContext function must not be null");
+        InjectableBean<T> bean = (InjectableBean<T>) contextual;
+        if (!Scopes.scopeMatches(this, bean)) {
+            throw Scopes.scopeDoesNotMatchException(this, bean);
+        }
+        CurrentContextState state = currentState();
+        if (state == null || !state.isValid()) {
+            return null;
+        }
+        ContextInstances contextInstances = state.contextInstances;
+        ContextInstanceHandle<T> instance = (ContextInstanceHandle<T>) contextInstances.getIfPresent(bean.getIdentifier());
+        if (instance == null) {
+            CreationalContext<T> creationalContext = creationalContextFun.apply(contextual);
+            return (T) contextInstances.computeIfAbsent(bean.getIdentifier(), new Supplier<ContextInstanceHandle<?>>() {
+
+                @Override
+                public ContextInstanceHandle<?> get() {
+                    return new ContextInstanceHandleImpl<>(bean, contextual.create(creationalContext), creationalContext);
+                }
+            }).get();
+        }
+        return instance.get();
+    }
+
+    @Override
+    public <T> T get(Contextual<T> contextual, CreationalContext<T> creationalContext) {
+        T result = getIfActive(contextual,
+                CreationalContextImpl.unwrap(Objects.requireNonNull(creationalContext, "CreationalContext must not be null")));
+        if (result == null) {
+            throw notActive();
+        }
+        return result;
+    }
+
+    @Override
+    public <T> T get(Contextual<T> contextual) {
+        Objects.requireNonNull(contextual, "Contextual must not be null");
+        InjectableBean<T> bean = (InjectableBean<T>) contextual;
+        if (!Scopes.scopeMatches(this, bean)) {
+            throw Scopes.scopeDoesNotMatchException(this, bean);
+        }
+        CurrentContextState state = currentState();
+        if (state == null || !state.isValid()) {
+            throw notActive();
+        }
+        @SuppressWarnings("unchecked")
+        ContextInstanceHandle<T> instance = (ContextInstanceHandle<T>) state.contextInstances
+                .getIfPresent(bean.getIdentifier());
+        return instance == null ? null : instance.get();
+    }
+
+    @Override
+    public boolean isActive() {
+        CurrentContextState contextState = currentState();
+        return contextState == null ? false : contextState.isValid();
+    }
+
+    @Override
+    public void destroy() {
+        destroy(currentState());
+    }
+
+    @Override
+    public void destroy(Contextual<?> contextual) {
+        CurrentContextState state = currentState();
+        if (state == null || !state.isValid()) {
+            throw notActive();
+        }
+        InjectableBean<?> bean = (InjectableBean<?>) contextual;
+        ContextInstanceHandle<?> instance = state.contextInstances.remove(bean.getIdentifier());
+        if (instance != null) {
+            instance.destroy();
+        }
+    }
+
+    @Override
+    public void destroy(ContextState state) {
+        if (traceLog().isTraceEnabled()) {
+            traceDestroy(state);
+        }
+        if (state == null) {
+            // nothing to destroy
+            return;
+        }
+        if (state instanceof CurrentContextState) {
+            CurrentContextState currentState = ((CurrentContextState) state);
+            if (currentState.invalidate()) {
+                fireIfNotNull(beforeDestroyedNotifier);
+                currentState.contextInstances.removeEach(ContextInstanceHandle::destroy);
+                fireIfNotNull(destroyedNotifier);
+            }
+        } else {
+            throw new IllegalArgumentException("Invalid state implementation: " + state.getClass().getName());
+        }
+    }
+
+    public CurrentContextState initializeContextState() {
+        CurrentContextState state = new CurrentContextState(contextInstances.get());
+        fireIfNotNull(initializedNotifier);
+        return state;
+    }
+
+    protected Logger traceLog() {
+        return LOG;
+    }
+
+    protected void traceActivate(ContextState initialState) {
+        // Noop
+    }
+
+    protected void traceDeactivate() {
+        // Noop
+    }
+
+    protected void traceDestroy(ContextState state) {
+        // Noop
+    }
+
+    private CurrentContextState currentState() {
+        return currentContext.get();
+    }
+
+    protected abstract ContextNotActiveException notActive();
+
+    private void fireIfNotNull(Consumer<Object> notifier) {
+        if (notifier != null) {
+            try {
+                notifier.accept(toString());
+            } catch (Exception e) {
+                LOG.warn("An error occurred during delivery of the context lifecycle event for " + toString(), e);
+            }
+        }
+    }
+
+    public static class CurrentContextState implements ContextState {
+
+        // Using 0 as default value enable removing an initialization
+        // in the constructor, piggybacking on the default value.
+        // As per https://docs.oracle.com/javase/specs/jls/se8/html/jls-12.html#jls-12.5
+        // the default field values are set before 'this' is accessible, hence
+        // they should be the very first value observable even in presence of
+        // unsafe publication of this object.
+        private static final int VALID = 0;
+        private static final int INVALID = 1;
+        private static final VarHandle IS_VALID;
+
+        static {
+            try {
+                IS_VALID = MethodHandles.lookup().findVarHandle(CurrentContextState.class, "isValid", int.class);
+            } catch (ReflectiveOperationException e) {
+                throw new Error(e);
+            }
+        }
+
+        private final ContextInstances contextInstances;
+        private volatile int isValid;
+
+        CurrentContextState(ContextInstances contextInstances) {
+            this.contextInstances = Objects.requireNonNull(contextInstances);
+        }
+
+        @Override
+        public Map<InjectableBean<?>, Object> getContextualInstances() {
+            return contextInstances.getAllPresent().stream()
+                    .collect(Collectors.toUnmodifiableMap(ContextInstanceHandle::getBean, ContextInstanceHandle::get));
+        }
+
+        /**
+         * @return {@code true} if the state was successfully invalidated, {@code false} otherwise
+         */
+        boolean invalidate() {
+            // Atomically sets the value just like AtomicBoolean.compareAndSet(boolean, boolean)
+            return IS_VALID.compareAndSet(this, VALID, INVALID);
+        }
+
+        @Override
+        public boolean isValid() {
+            return isValid == VALID;
+        }
+
+    }
+
+}
